fix(pay): stop polling expired FamPay QRs; keep last-seen precise
Sweep was 408-ing dead orders every minute, so Check status looked like
a blip and All users last-seen rounded to hours on Postgres REAL.

+++ src/uotpbot/wallets.py
@@ -326,8 +326,8 @@ _SEEN_SCHEMA = """
 CREATE TABLE IF NOT EXISTS {t} (
     scope TEXT NOT NULL DEFAULT '',
     user_id TEXT NOT NULL,
-    first_seen REAL NOT NULL,
-    last_seen REAL NOT NULL,
+    first_seen DOUBLE PRECISION NOT NULL,
+    last_seen DOUBLE PRECISION NOT NULL,
     username TEXT NOT NULL DEFAULT '',
     PRIMARY KEY (scope, user_id)
 )
